refactor: 각 기물들의 이동 전략을 MoveStrategy로 추상화하고 Piece를 구체 클레스로 변경
