+++ src/main/java/domain/Moves.java
@@ -1,8 +1,10 @@
 package domain;
 
+import domain.piece.Position;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 
 public class Moves {
 
@@ -52,5 +54,22 @@ public class Moves {
             positionInPath = positionInPath.movePosition(move);
         }
         return true;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        Moves moves1 = (Moves) o;
+        return Objects.equals(moves, moves1.moves);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hashCode(moves);
     }
 }
