--- src/test/java/domain/piece/PoTest.java
@@ -1,136 +1,0 @@
-package domain.piece;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-import domain.Position;
-import domain.Team;
-import java.util.List;
-import java.util.stream.Stream;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.MethodSource;
-
-class PoTest {
-
-    @Test
-    void 포가_위로_이동하는_경로를_계산할_수_있다() {
-        // given
-        Position src = new Position(4, 1);
-        Position dest = new Position(1, 1);
-        Po po = new Po(Team.HAN, src);
-
-        // when
-        List<Position> path = po.calculatePath(src, dest);
-
-        // then
-        List<Position> expected = List.of(new Position(3, 1), new Position(2, 1));
-        assertThat(path).isEqualTo(expected);
-    }
-
-    @Test
-    void 포가_아래로_이동하는_경로를_계산할_수_있다() {
-        // given
-        Position src = new Position(1, 1);
-        Position dest = new Position(4, 1);
-        Po po = new Po(Team.HAN, src);
-
-        // when
-        List<Position> path = po.calculatePath(src, dest);
-
-        // then
-        List<Position> expected = List.of(new Position(2, 1), new Position(3, 1));
-        assertThat(path).isEqualTo(expected);
-    }
-
-    @Test
-    void 포가_왼쪽으로_이동하는_경로를_계산할_수_있다() {
-        // given
-        Position src = new Position(4, 4);
-        Position dest = new Position(1, 4);
-        Po po = new Po(Team.HAN, src);
-
-        // when
-        List<Position> path = po.calculatePath(src, dest);
-
-        // then
-        List<Position> expected = List.of(new Position(3, 4), new Position(2, 4));
-        assertThat(path).isEqualTo(expected);
-    }
-
-    @Test
-    void 포가_오른쪽으로_이동하는_경로를_계산할_수_있다() {
-        // given
-        Position src = new Position(1, 1);
-        Position dest = new Position(1, 4);
-        Po po = new Po(Team.HAN, src);
-
-        // when
-        List<Position> path1 = po.calculatePath(src, dest);
-
-        // then
-        List<Position> expected1 = List.of(new Position(1, 2), new Position(1, 3));
-        assertThat(path1).isEqualTo(expected1);
-    }
-
-    @ParameterizedTest
-    @CsvSource({
-            "1, 1",
-            "1, -1",
-            "-1, 1",
-            "-1, -1"
-    })
-    void 포로_이동할_수_없는_위치인_경우_예외를_발생시킨다(int movedRow, int movedColumn) {
-        // given
-        int row = 5;
-        int column = 5;
-        Position src = new Position(row, column);
-        Po po = new Po(Team.HAN, src);
-        Position dest = new Position(row + movedRow, column + movedColumn);
-        // when & then
-        assertThatThrownBy(() -> po.calculatePath(src, dest))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("이 위치로 이동할 수 없습니다.");
-    }
-
-    @ParameterizedTest
-    @MethodSource(value = "createDiagonalMovePosition")
-    void 포가_궁성의_꼭짓점에서_꼭짓점으로_이동하는_경로를_계산할_수_있다(Position src, Position dest, Position jumpedPosition) {
-        // given
-        Po po = new Po(Team.HAN, src);
-
-        // when
-        List<Position> path = po.calculatePath(src, dest);
-
-        // then
-        List<Position> expected = List.of(jumpedPosition);
-        assertThat(path).isEqualTo(expected);
-    }
-
-    private static Stream<Arguments> createDiagonalMovePosition() {
-        return Stream.of(
-                Arguments.of(
-                        new Position(1, 4),
-                        new Position(3, 6),
-                        new Position(2, 5)
-                ),
-                Arguments.of(
-                        new Position(1, 6),
-                        new Position(3, 4),
-                        new Position(2, 5)
-                ),
-                Arguments.of(
-                        new Position(3, 4),
-                        new Position(1, 6),
-                        new Position(2, 5)
-                ),
-                Arguments.of(
-                        new Position(3, 6),
-                        new Position(1, 4),
-                        new Position(2, 5)
-                )
-        );
-    }
-}
